Before: all ci must pass in pr without any warnings or failings
